% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdCarbon.R
\name{calc_cwd_c}
\alias{calc_cwd_c}
\title{Coarse Woody Debris Carbon}
\usage{
calc_cwd_c(volume_ha, Decay_class, Species, BECzone = "SBS")
}
\arguments{
\item{volume_ha}{volume/ha of CWD piece or plot}

\item{Decay_class}{Decay class of CWD piece or plot}

\item{Species}{Species of CWD piece or plot}
}
\description{
This function calculates the amount of carbon in a piece or plot of coarse woody debris based
on the species and decay class. Data is passed to the function by species and decay class. The function will
estimate the carbon of unknown species based on the BEC zone (currently SBS or ICH only, default is set to SBS).

Currently, the function estimates carbon for: Lodgepole pine, hybrid spruce, subalpine fir, trembling aspen,
paper birch, red alder, cottonwood, amabilis fir, western hemlock, western redcedar, Douglas-fir and
western larch
}
\details{
cwdCARBON (Mg/ha) = volume(m3/ha)
x structural reduction factor (decay class specific (Fraver et al. 2013))
x Absolute density(g/cm3) (species and decay class specific (Harmon et al. 2008))
x CarbonConcentration (species and decay class specific (Harmon et al. 2013))
}
