% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdCarbon.R
\name{calc_fwd_c}
\alias{calc_fwd_c}
\title{Calculate fine woody debris carbon}
\usage{
calc_fwd_c(Diam_class, volume)
}
\arguments{
\item{volume}{fwdCARBON (Mg/ha) = volume(m3/ha)
x Live wood density(g/cm3) # use unknown species (Harmon et al. 2008)
x Decay reduction factor for each size class (Harmon and Fasth website)
x CarbonConcentration
use 50\% (Harmon and Fasth website)}
}
\description{
Calculate fine woody debris carbon
}
