% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallTreeCarbon.R
\name{calc_sm_tree_c_Annigofer}
\alias{calc_sm_tree_c_Annigofer}
\title{Small tree Annigofer equations}
\usage{
calc_sm_tree_c_Annigofer(Species, Height_class)
}
\arguments{
\item{Height_class}{}
}
\description{
Small tree Annigofer equations
}
