% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallTreeCarbon.R
\name{calc_sm_tree_c_Ung}
\alias{calc_sm_tree_c_Ung}
\title{calculate small tree carbon}
\usage{
calc_sm_tree_c_Ung(Species, Height_class, Diam_est, Health)
}
\arguments{
\item{Species}{current valid species are At, Ac, Cw, Bl, Ep, Hw, Pl, Sx, Fd, UC, Lw}

\item{Health}{}
}
\description{
calculate small tree carbon
}
\details{
Live_Dead - should we specify live and dead? Is it worth it?
}
