% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_cshapley}
\alias{cpp_cshapley}
\title{Compute hierarchical cShapley values for tree leaves}
\usage{
cpp_cshapley(tree_list, calc_func)
}
\arguments{
\item{tree_list}{A nested list representing the tree hierarchy.}

\item{calc_func}{An R function taking a DataFrame of leaves with columns 'key' and 'value'.}
}
\value{
A named numeric vector of Shapley values for each leaf.
}
\description{
Compute hierarchical cShapley values for tree leaves
}
