% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_permute_tree}
\alias{cpp_permute_tree}
\title{Permute all subtrees of a tree in C++}
\usage{
cpp_permute_tree(tree_list)
}
\arguments{
\item{tree_list}{A nested list representing the tree hierarchy.}
}
\value{
A list of nested lists representing all permuted trees.
}
\description{
Generates all valid permutations of children at each internal node level.
}
