% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_tree_to_string}
\alias{cpp_tree_to_string}
\title{Generate ASCII visual representation of a tree}
\usage{
cpp_tree_to_string(tree_list)
}
\arguments{
\item{tree_list}{A nested list representing the tree hierarchy.}
}
\value{
A character string showing the ASCII tree.
}
\description{
Generate ASCII visual representation of a tree
}
