% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdCarbon.R
\name{cwdCarbonFN}
\alias{cwdCarbonFN}
\title{Coarse Woody Debris Carbon}
\usage{
cwdCarbonFN(volume_ha, Decay_class, Species, BECzone = "SBS")
}
\arguments{
\item{Species}{}
}
\description{
Coarse Woody Debris Carbon
}
\details{
cwdCARBON (Mg/ha) = volume(m3/ha)
x structural reduction factor # decay class specific (Fraver et al. 2013)
x Absolute density(g/cm3) # species and decay class specific (Harmon et al. 2008)
x CarbonConcentration # species and decay class specific (Harmon et al. 2013)
}
