% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdCarbon.R
\name{cwdCfromSORTIE}
\alias{cwdCfromSORTIE}
\title{Title}
\usage{
cwdCfromSORTIE(VolGrid, speciesGroupings, SizeClassCut)
}
\arguments{
\item{VolGrid}{the table of volumes from SORTIE}

\item{speciesGroupings}{data.frame of species and groups}

\item{SizeClassCut}{cutoff between small and large logs (20 right now)}
}
\description{
Title
}
