% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeDiamHgtAllometry.R
\name{diam_hgt_summitLake}
\alias{diam_hgt_summitLake}
\title{Summit Lake Diameter-Height relationships}
\usage{
diam_hgt_summitLake(DBH)
}
\arguments{
\item{DBH}{}
}
\description{
Summit Lake Diameter-Height relationships
}
