% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeDiamHgtAllometry.R
\name{DiamHgtFN}
\alias{DiamHgtFN}
\title{Diameter-Height functions}
\usage{
DiamHgtFN(Species, DBH, BECzone = "SBS")
}
\description{
The "standard" diameter-height relationship from SORTIE.
The parameter values were extracted from the SBS and ICH parameter files
}
\details{
allometry varies with environment, so we use the paramter estimates
based on BEC zones for these functions and assigned a default of the SBS:

From ICH parameter file:
Slope of asymptotic height:
"Western_Hemlock" 0.0299,"Western_redcedar" 0.0241,
"Amabalis_Fir" 0.0263,"Subalpine_Fir" 0.0264,
"Hybrid_spruce" 0.0264,"Lodgepole_Pine" 0.0333,
"Trembling_Aspen" 0.0352, "Black_Cottonwood" 0.0347,
"Paper_Birch" 0.0454

Canopy tree height:
"Western_Hemlock" 39.48, "Western_redcedar" 39.54,
"Amabalis_Fir" 40.0, "Subalpine_Fir" 40.0,
"Hybrid_spruce" 45.0, "Lodgepole_Pine" 40.0,
"Trembling_Aspen" 39.14, "Black_Cottonwood" 39.47,
"Paper_Birch" 33.18

From the SBS parameter file:
"Interior_Spruce" 0.0299364, "Lodgepole_Pine" 0.070728,
"Subalpine_Fir" 0.0305068, "Trembling_Aspen" 0.03746

Canopy tree height:
"Interior_Spruce" 35.0, "Lodgepole_Pine" 23.346836,
"Subalpine_Fir" 30.0, "Trembling_Aspen" 33.53
}
