% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_interface.R
\name{from_tree_list}
\alias{from_tree_list}
\title{Convert a nested tree list into a data.tree Node (if data.tree is available)}
\usage{
from_tree_list(l, as_data_tree = NULL)
}
\arguments{
\item{l}{A nested list.}

\item{as_data_tree}{Logical; if TRUE, converts to data.tree Node. If FALSE, returns list.
If NULL (default), converts to data.tree if package is installed.}
}
\value{
A `data.tree` Node or nested list.
}
\description{
Convert a nested tree list into a data.tree Node (if data.tree is available)
}
