% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeDiamHgtAllometry.R
\name{height_dbh_plantations}
\alias{height_dbh_plantations}
\title{Diam - height function in clearcuts (high light)}
\usage{
height_dbh_plantations(Species, DBH, BECzone = "ICH")
}
\arguments{
\item{BECzone}{only ICH supported}
}
\description{
Diam - height function in clearcuts (high light)
}
\details{
estimating allometry in a high light environment (clearcut). If estimating residual trees
(e.g  those that remain after a partial harvest) please see
height_dbh_residual function instead. Both of these
functions only apply to the ICH.
}
