% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeDiamHgtAllometry.R
\name{height_dbh_Residuals}
\alias{height_dbh_Residuals}
\title{Diam-height function residuals}
\usage{
height_dbh_Residuals(Species, DBH, BECzone = "ICH")
}
\arguments{
\item{BECzone}{}
}
\description{
Diam-height function residuals
}
