% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_interface.R
\name{is_datatree_node}
\alias{is_datatree_node}
\title{Check if an object is a data.tree Node}
\usage{
is_datatree_node(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
Logical indicating if x inherits from 'Node'.
}
\description{
Check if an object is a data.tree Node
}
\keyword{internal}
