% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_interface.R
\name{print_tree}
\alias{print_tree}
\title{Print ASCII visual representation of a tree}
\usage{
print_tree(tree)
}
\arguments{
\item{tree}{A `data.tree` Node or nested tree list.}
}
\value{
Invisibly returns the formatted string while printing to console.
}
\description{
Print ASCII visual representation of a tree
}
