% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortie_carbon.R
\name{read_sortie}
\alias{read_sortie}
\title{read SORTIE}
\usage{
read_sortie(outputs)
}
\arguments{
\item{outputs}{expecting a single csv for each site/plot with all timesteps or a vector of
filenames to read in}
}
\description{
read SORTIE
}
