% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortie_carbon.R
\name{sortie_cwd_carbon}
\alias{sortie_cwd_carbon}
\title{Coarse woody debris carbon}
\usage{
sortie_cwd_carbon(volGrid, spGroups)
}
\arguments{
\item{spGroups}{}
}
\description{
To do - make a default species group table and throw errors if species don't match
}
