% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortie_carbon.R
\name{sortie_tree_carbon}
\alias{sortie_tree_carbon}
\title{Title}
\usage{
sortie_tree_carbon(sortie_outputs, BEC, ht_from_diam = "standard")
}
\arguments{
\item{sortie_outputs}{sortie output saved in a data.table}

\item{BEC}{Biogeoclimatic ecosystem classification zone}

\item{dead}{include snags in the calculation}

\item{Ht_from_diam}{calculate the height based on the diameter (allometry) standard
plantation or residual are all options (ICH only for plantation or residual). "none" means
use whichever height is already there.}
}
\description{
Title
}
\details{
Currently only supports common species in the SBS and ICH
To do: add argument for seedling to call the seedling carbon function
}
