% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stubs.R
\name{StubVol_ConicPara}
\alias{StubVol_ConicPara}
\title{Title}
\usage{
StubVol_ConicPara(Height, DBH, Species, StubHT)
}
\arguments{
\item{StubHT}{}
}
\value{
returns the estimate of volume/ha (not sure how it becomes per hectare - check w Erica)
}
\description{
Title
}
