% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_interface.R
\name{traverse_tree}
\alias{traverse_tree}
\title{Traverse a tree and return leaf keys and values}
\usage{
traverse_tree(tree)
}
\arguments{
\item{tree}{A `data.tree` Node or nested tree list.}
}
\value{
A `data.frame` containing leaf keys and values in traversal order.
}
\description{
Traverse a tree and return leaf keys and values
}
