% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_interface.R
\name{tree_leaf}
\alias{tree_leaf}
\title{Helper to create a leaf node list}
\usage{
tree_leaf(name, value = 0)
}
\arguments{
\item{name}{Character string; key/name of the leaf node.}

\item{value}{Numeric value stored in the leaf.}
}
\value{
A list representing a leaf node.
}
\description{
Helper to create a leaf node list
}
