% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_interface.R
\name{tree_node}
\alias{tree_node}
\title{Helper to create an internal node list}
\usage{
tree_node(name, children = list(), value = NULL)
}
\arguments{
\item{name}{Character string; key/name of the internal node.}

\item{children}{List of child nodes.}

\item{value}{Optional numeric value for the internal node.}
}
\value{
A list representing an internal node.
}
\description{
Helper to create an internal node list
}
