\name{TreeCalcs-package}
\alias{TreeCalcs-package}
\alias{TreeCalcs}
\docType{package}
\title{Fast Tree Permutations, Traversals, and Aggregations in C++20}
\description{
Efficient tree algorithms powered by C++20 concepts and Rcpp,
supporting native data.tree Node objects and standard nested R lists.
}
\author{
Tim \email{tim@example.com}
}
\keyword{package}
