% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merchTreeVolume.R
\name{TreeMerchVolFN}
\alias{TreeMerchVolFN}
\title{Merchantable Timber Volume}
\usage{
TreeMerchVolFN(Species, DBH, HT, BECzone = "SBS")
}
\arguments{
\item{Species}{tree species}

\item{DBH}{diameter (cm) at breast height (DBH) of an indovidual tree}

\item{HT}{total tree height (m) of an individual tree}

\item{BECzone}{Biogeoclimatic zone. Currently only sub boreal spruce (SBS) and interior cedar hemlock (ICH)}
}
\description{
This function calculates the volume (m^3) of merchantable timber. Data is passed to the function by species and BECzone.
The function will estimate the merchantable volume of unknown species based on the BEC zone (currently SBS or ICH only, default is set to SBS).

Currently, the function estimates volume for: Lodgepole pine, hybrid spruce, subalpine fir, trembling aspen,
paper birch, red alder, cottonwood, amabilis fir, western hemlock, western redcedar, Douglas-fir and
western larch
}
\details{
merchantable timber volumber (m^3) = e^b0 x DBH^b1 x HT^b2
where b0, b1, b2 are model parameters from North et al. 2016 Total and merchantable volume equations for common tree species in BC
}
