% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeDiamHgtAllometry.R
\name{height_dbh_plantations}
\alias{height_dbh_plantations}
\title{Diam - height function in clearcuts (high light)}
\usage{
height_dbh_plantations(Species, DBH, BECzone = "ICH")
}
\arguments{
\item{BECzone}{}
}
\description{
Diam - height function in clearcuts (high light)
}
