% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeDiamHgtAllometry.R
\name{height_dbh_Residuals}
\alias{height_dbh_Residuals}
\title{Diam-height function residuals}
\usage{
height_dbh_Residuals(Species, DBH, BECzone = "ICH")
}
\arguments{
\item{BECzone}{only ICH supported}
}
\description{
Diam-height function residuals
}
\details{
estimating allometry on residual trees (e.g  those that remain after a partial harvest)
if trees are growing in a high light environment (clearcut) in the ICH, please see
height_dbh_plantations function instead. Both of these
functions only apply to the ICH.
}
